# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ma premiere api laravel",
        "description": "This is a sample server Petstore server.  You can find out more about Swagger at <a href=\"http://swagger.io\">http://swagger.io</a> or on irc.freenode.net, #swagger.  For this sample, you can use the api key \"special-key\" to test the authorization filters",
        "termsOfService": "http://helloreverb.com/terms/",
        "contact": {
            "email": "airskual@yahoo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "myapi.com",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/film": {
            "get": {
                "tags": [
                    "film"
                ],
                "summary": "Display a list of films.",
                "description": "This can only be done by the logged in user.",
                "operationId": "getFilm",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Film"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "film"
                ],
                "summary": "add 1 film.",
                "description": "This is to insert a film",
                "operationId": "addFilm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "titre",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "id_genre",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "id_distributeur",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "resum",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "date_debut_affiche",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "date_fin_affiche",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "duree_minutes",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "annee_production",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Film"
                            }
                        }
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/film/{filmId}": {
            "get": {
                "tags": [
                    "film"
                ],
                "summary": "show 1 row",
                "description": "Show one row",
                "operationId": "getFilmById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "filmId",
                        "in": "path",
                        "description": "ID of film that needs to be fetched",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Film"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "film"
                ],
                "summary": "Update an existing film",
                "description": "",
                "operationId": "updateFilm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "filmId",
                        "in": "path",
                        "description": "Pet object that needs to be added to the store",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "titre",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "id_genre",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "id_distributeur",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "resum",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "date_debut_affiche",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "date_fin_affiche",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "duree_minutes",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "annee_production",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Invalid ID supplied",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Film"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Film not found"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:films",
                            "read:films"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "film"
                ],
                "summary": "Delete purchase order by ID",
                "description": "For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors",
                "operationId": "deleteFilm",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "filmId",
                        "in": "path",
                        "description": "ID of the order that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/genre": {
            "get": {
                "tags": [
                    "genre"
                ],
                "summary": "Display a list of films.",
                "description": "This can only be done by the logged in user.",
                "operationId": "getGenre",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Film"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "genre"
                ],
                "summary": "add 1 film.",
                "description": "This is to insert a genre",
                "operationId": "addGenre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "nom",
                        "in": "formData",
                        "description": "the name of the field you want to update",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Film"
                            }
                        }
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/genre/{genreId}": {
            "get": {
                "tags": [
                    "genre"
                ],
                "summary": "show 1 row",
                "description": "Show one row",
                "operationId": "getGenreById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "genreId",
                        "in": "path",
                        "description": "ID of genre that needs to be fetched",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Genre"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "genre"
                ],
                "summary": "Delete genre by ID",
                "operationId": "deleteGenre",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "genreId",
                        "in": "path",
                        "description": "ID of the genre that needs to be deleted",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/film/{genreId}": {
            "put": {
                "tags": [
                    "genre"
                ],
                "summary": "Update an existing genre",
                "description": "",
                "operationId": "updateGenre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "genreId",
                        "in": "path",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "nom",
                        "in": "formData",
                        "description": "the fields you want to update",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Invalid ID supplied",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Film"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Film not found"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:films",
                            "read:films"
                        ]
                    }
                ]
            }
        }
    },
    "definitions": {
        "Film": {
            "required": true,
            "properties": {
                "id_film": {
                    "type": "integer",
                    "format": "int64",
                    "default": 4554
                },
                "titre": {
                    "type": "string"
                },
                "resum": {
                    "type": "string"
                },
                "date_debut_affiche": {
                    "type": "date",
                    "default": "2011-01-12"
                },
                "date_fin_affiche": {
                    "type": "date",
                    "format": "date",
                    "default": "2011-01-12"
                },
                "duree_minutes": {
                    "type": "integer",
                    "default": 180
                },
                "annee_production": {
                    "type": "integer",
                    "default": 2011
                }
            },
            "xml": {
                "name": "Film"
            }
        },
        "Genre": {
            "required": true,
            "properties": {
                "id_film": {
                    "type": "integer",
                    "format": "int64",
                    "default": 4554
                },
                "titre": {
                    "type": "string",
                    "default": "Erotic"
                }
            },
            "xml": {
                "name": "Genre"
            }
        }
    }
}